Consolidate discount entity files
