--- src/resources/entities/discount/ISimpleDiscount.ts
@@ -1,19 +1,0 @@
-export default interface SimpleDiscount {
-
-    cik: string;
-
-    symbol: string;
-
-    name: string;
-
-    active: boolean;
-
-    ratioPrice?: number;
-
-    ttmSalePrice: number;
-
-    tfySalePrice: number;
-
-    ttySalePrice: number;
-    
-}
