--- src/resources/entities/discount/models/TrailingPriceData.ts
@@ -1,9 +1,0 @@
-export default interface TrailingPriceData {
-
-    cik: string;
-
-    stickerPrice: number;
-
-    salePrice: number;
-
-}
